+++ crates/language/Cargo.toml
@@ -68,6 +68,7 @@ util.workspace = true
 watch.workspace = true
 zlog.workspace = true
 diffy = "0.4.2"
+encoding = "0.2.33"
 
 [dev-dependencies]
 collections = { workspace = true, features = ["test-support"] }
