+++ crates/zed/src/zed.rs
@@ -443,7 +443,10 @@ pub fn initialize_workspace(
             }
         });
 
-        let encoding_indicator = cx.new(|_cx| encodings::Encoding::Utf8(workspace.weak_handle()));
+        let encoding_indicator = cx.new(|_cx| encodings::EncodingIndicator {
+            encoding: encodings::Encoding::Utf8,
+            workspace: workspace_handle.downgrade(),
+        });
 
         let cursor_position =
             cx.new(|_| go_to_line::cursor_position::CursorPosition::new(workspace));
